# api/swagger.yaml
openapi: 3.0.1
info:
  title: Simple Bookstore
  description: >-
    This is the bookstore server  for the Hypermedia project course 2019. It
    contains a book inventory, users and
     a shopping cart.
  contact:
    email: piervincenzov@gmail.com
  license:
    name: Apache-2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: "1.0.0-oas3"
...
servers:
  - url: 'https://ventra-books.herokuapp.com/'
...
paths:
...
  /user/login:
    post:
      tags:
        - user
      summary: Login
      description: Login with a form
      operationId: userLoginPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
        required: true
      responses:
        '200':
          description: succesfull login
        '400':
          description: Wrong password or email, or duplicated session id on the server.
        '500':
          description: Internal server error.
      x-swagger-router-controller: User
  /user/logout:
    post:
      tags:
        - user
      summary: Logout
      description: Logout request
      operationId: userLogoutPOST
      responses:
        '200':
          description: succesful logout
        '401':
          description: Unauthorized
        '500':
          description: Internal server error.
      x-swagger-router-controller: User
  /user/register:
    post:
      tags:
        - user
      summary: Create a new user.
      description: Register into the store.
      operationId: userRegisterPOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterBody'
      responses:
        '200':
          description: succesfull login
        '400':
          description: Email already in use, or empty fields.
        '500':
          description: Internal server error.
      x-swagger-router-controller: User
  '/user/{userId}/userPurchases':
    get:
      tags:
        - user
      summary: retrieve all the purchases of the given user
      description: returns a list of Purchases
      operationId: getUserPurchases
      parameters:
        - name: userId
          in: path
          description: Id of the User
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A collection of Purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Purchase'
                x-content-type: application/json
      security:
        - cookieAuth: []
      x-swagger-router-controller: User
  /user/cart:
    get:
      tags:
        - user
      summary: retrieve the books contained in the given Cart
      description: returns a list of Cart Items
      operationId: getCart
      responses:
        '200':
          description: A collection of Books contained in the Cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
                x-content-type: application/json
        '401':
          description: Unauthenticated user.
      security:
        - cookieAuth: []
      x-swagger-router-controller: User
components:
  schemas:
...
    Purchase:
      title: Purchase
      required:
        - user_id
        - date_time
        - book
        - amount
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        date_time:
          type: string
        book:
          $ref: '#/components/schemas/BookPreview'
        amount:
          type: string
      description: A purchase that the user performed
      example:
        id: 0
        user_id: 1
        date_time: 12.03.18
        book:
          isbn: 9788817153096
          price: eur 6.80
          title: The Adventures of Huckleberry Finn
          format: paper
          picture: /assets/img/9788817153096
        amount: eur 6.80
    BookPreview:
      title: BookPreview
      required:
        - isbn
        - price
        - title
        - format
        - picture
      type: object
      properties:
        isbn:
          type: string
        price:
          type: string
        title:
          type: string
        format:
          type: string
        picture:
          type: string
      description: 'A preview of a Book, containing only the essential info'
      example:
        isbn: 9788817153096
        price: eur 6.80
        title: The Adventures of Huckleberry Finn
        format: paper
        picture: /assets/img/9788817153096
...
    CartItem:
      title: CartItem
      type: object
      properties:
        book:
          $ref: '#/components/schemas/BookPreview'
        quantity:
          type: integer
    LoginBody:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    RegisterBody:
      required:
        - email
        - firstname
        - lastname
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstname:
          type: string
        lastname:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: JSESSIONID
      in: cookie